Extract scripts into separate file
